finish !

+++ src/main/java/cz/jan/order/model/OrderItem.java
@@ -1,9 +1,11 @@
 package cz.jan.order.model;
 
+import java.math.BigDecimal;
+
 public record OrderItem(
         Long productId,
         String name,
         Long quantity,
-        Double pricePerUnit
+        BigDecimal pricePerUnit
 ) {
 }
