+++ src/main/java/cz/jan/product/model/Product.java
@@ -1,10 +1,12 @@
 package cz.jan.product.model;
 
+import java.math.BigDecimal;
+
 public record Product(
         Long id,
         String name,
         Long quantity,
-        Double pricePerUnit,
+        BigDecimal pricePerUnit,
         Boolean active
 ) {
 }
